feat(forecast): average change
